refactor: internal/files
