+++ deb/deb.go
@@ -16,6 +16,7 @@ import (
 	"time"
 
 	"github.com/blakesmith/ar"
+	"github.com/goreleaser/nfpm/internal/files"
 	"github.com/pkg/errors"
 
 	"github.com/goreleaser/chglog"
@@ -170,11 +171,11 @@ func createFilesInsideTarGz(info *nfpm.Info, out *tar.Writer, created map[string
 	var md5buf bytes.Buffer
 	var instSize int64
 
-	files, err := info.FilesToCopy()
+	copiable, err := files.FilesToCopy(info)
 	if err != nil {
 		return md5buf, 0, err
 	}
-	for _, file := range files {
+	for _, file := range copiable {
 		if err = createTree(out, file.Destination, created); err != nil {
 			return md5buf, 0, err
 		}
