--- internal/helpers/helpers.go
@@ -1,10 +1,0 @@
-// Package helpers provides common helper methods
-package helpers
-
-import "strconv"
-
-// IsInt returns true if s is int, false otherwise.
-func IsInt(s string) bool {
-	_, err := strconv.Atoi(s)
-	return err == nil
-}
