+++ nfpm.go
@@ -8,19 +8,13 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"sort"
-	"strings"
+	"strconv"
 	"sync"
 
 	"github.com/Masterminds/semver/v3"
+	"github.com/goreleaser/chglog"
 	"github.com/imdario/mergo"
 	"github.com/pkg/errors"
-
-	"github.com/goreleaser/chglog"
-
-	"github.com/goreleaser/nfpm/glob"
-	"github.com/goreleaser/nfpm/internal/helpers"
-
 	"gopkg.in/yaml.v2"
 )
 
@@ -234,56 +228,15 @@ func WithDefaults(info *Info) *Info {
 	// and support proper ordering for both rpm and deb
 	if v, err := semver.NewVersion(info.Version); err == nil {
 		info.Version = fmt.Sprintf("%d.%d.%d", v.Major(), v.Minor(), v.Patch())
-		if info.Release == "" && helpers.IsInt(v.Prerelease()) {
+		if info.Release == "" && isInt(v.Prerelease()) {
 			info.Release = v.Prerelease()
 		}
-		if info.Prerelease == "" && !helpers.IsInt(v.Prerelease()) {
+		if info.Prerelease == "" && !isInt(v.Prerelease()) {
 			info.Prerelease = v.Prerelease()
 		}
 	}
 
 	return info
-}
-
-// FileToCopy describes the source and destination of one file to copy into a
-// package and whether it is a config file.
-type FileToCopy struct {
-	Source      string
-	Destination string
-	Config      bool
-}
-
-// FilesToCopy lists all of the real files to be copied into the package.
-func (info *Info) FilesToCopy() ([]FileToCopy, error) {
-	var files []FileToCopy
-	for i, filesMap := range []map[string]string{info.Files, info.ConfigFiles} {
-		for srcglob, dstroot := range filesMap {
-			globbed, err := glob.Glob(srcglob, dstroot)
-			if err != nil {
-				return nil, err
-			}
-			for src, dst := range globbed {
-				// avoid including a partial file with the name of the target in the target
-				// itself. when used as a lib, target may not be set. in that case, src will
-				// always have the empty sufix, and all files will be ignored.
-				if info.Target != "" && strings.HasSuffix(src, info.Target) {
-					fmt.Printf("skipping %s because it has the suffix %s", src, info.Target)
-					continue
-				}
-
-				files = append(files, FileToCopy{src, dst, i == 1})
-			}
-		}
-	}
-	// sort the files for reproducibility and general cleanliness
-	sort.Slice(files, func(i, j int) bool {
-		a, b := files[i], files[j]
-		if a.Source != b.Source {
-			return a.Source < b.Source
-		}
-		return a.Destination < b.Destination
-	})
-	return files, nil
 }
 
 // GetChangeLog parses the provided changelog file.
@@ -303,4 +256,9 @@ func (info *Info) GetChangeLog() (log *chglog.PackageChangeLog, err error) {
 		Name:    info.Name,
 		Entries: entries,
 	}, nil
+}
+
+func isInt(s string) bool {
+	_, err := strconv.Atoi(s)
+	return err == nil
 }
