+++ rpm/rpm.go
@@ -13,6 +13,7 @@ import (
 	"time"
 
 	"github.com/google/rpmpack"
+	"github.com/goreleaser/nfpm/internal/files"
 	"github.com/pkg/errors"
 	"github.com/sassoftware/go-rpmutils/cpio"
 
@@ -305,11 +306,11 @@ func addEmptyDirsRPM(info *nfpm.Info, rpm *rpmpack.RPM) {
 }
 
 func createFilesInsideRPM(info *nfpm.Info, rpm *rpmpack.RPM) error {
-	files, err := info.FilesToCopy()
+	copiable, err := files.FilesToCopy(info)
 	if err != nil {
 		return err
 	}
-	for _, file := range files {
+	for _, file := range copiable {
 		err := copyToRPM(rpm, file.Source, file.Destination, file.Config)
 		if err != nil {
 			return err
